feat: isolated the schemas creation logic
