+++ lib/src/functions/api/index.js
@@ -3,6 +3,7 @@ const util = require('util')
 const exec = util.promisify(require('child_process').exec)
 const writeFile = require('../../utils/writeFile')
 const database = require('./database')
+const schemas = require('./schemas')
 
 /*
  * Express api:
@@ -143,58 +144,6 @@ export {}
 export * from './server'
 `,
         file: `${projectName}/src/network/index.ts`
-      }
-    },
-    schemas: {
-      index: {
-        content: `import { Static, Type } from '@sinclair/typebox'
-
-const id = Type.String({
-  pattern: '^[a-zA-Z0-9]{24,}$'
-})
-
-type Id = Static<typeof id>
-
-const idSchema = Type.Object({ id })
-
-type IdSchema = Static<typeof idSchema>
-
-export { id, Id, idSchema, IdSchema }
-export * from './user'
-`,
-        file: `${projectName}/src/schemas/index.ts`
-      },
-      user: {
-        content: `import { Static, Type } from '@sinclair/typebox'
-
-import { id } from '.'
-
-const user = Type.Object({
-  lastName: Type.String(),
-  name: Type.String()
-})
-
-type User = Static<typeof user>
-
-const userDto = Type.Object({
-  id: Type.Optional(id),
-  lastName: Type.String(),
-  name: Type.String(),
-  createdAt: Type.Optional(Type.String()),
-  updatedAt: Type.Optional(Type.String())
-})
-
-type UserDTO = Static<typeof userDto>
-
-const storeUserSchema = Type.Object({
-  args: user
-})
-
-type StoreUser = Static<typeof storeUserSchema>
-
-export { userDto, UserDTO, user, User, storeUserSchema, StoreUser }
-`,
-        file: `${projectName}/src/schemas/user.ts`
       }
     },
     services: {
@@ -1738,7 +1687,6 @@ ${projectName}/src/@types \
 ${projectName}/src/@types/models \
 ${projectName}/src/network \
 ${projectName}/src/network/routes \
-${projectName}/src/schemas \
 ${projectName}/src/services \
 ${projectName}/src/services/utils \
 ${projectName}/src/services/utils/messages \
@@ -1754,6 +1702,8 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
     mongo,
     projectName
   })
+  // /schemas
+  schemas({ projectName })
 
   // /@types
   await writeFile(data['@types'].index.file, data['@types'].index.content)
@@ -1763,10 +1713,6 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
     data['@types/models'].user.file,
     data['@types/models'].user.content
   )
-
-  // /schemas
-  await writeFile(data.schemas.user.file, data.schemas.user.content)
-  await writeFile(data.schemas.index.file, data.schemas.index.content)
 
   // /services
   await writeFile(data.services.user.file, data.services.user.content)
